refactor(bridge): carve /usage /budget /pricing handlers (#79 Phase N)
Fourteenth carve. Three cost-reporting Telegram cmd handlers move to
`telegram_handlers/cost.py`. All deps (pricing.usage, budget.core,
budget.engine, pricing.snapshot, notify.telegram) live in carved
modules already, so this is mostly code motion.

Symbols moved (3):
- cmd_usage   — /usage    토큰 사용량 + 비용
- cmd_budget  — /budget   한도 설정 + 진척률 표시
- cmd_pricing — /pricing  LiteLLM 가격 스냅샷 + drift

bot.py wraps each via re-export so the dispatcher in `main()` keeps
referencing the same names.

bot.py wrapper for `take_pricing_snapshot` removed:
The wrapper existed to inject `send_telegram` into the pure
`pricing.snapshot.take_pricing_snapshot`. With cmd_pricing carved
out (which wired send_telegram inline), the only remaining caller
was `daily_pricing_snapshot` — also inlined the `send_alert=
send_telegram` arg. Net: -12 lines + a clearer call site.

cmd_pricing's `/pricing snapshot` and the daily scheduler now both
call `pricing.snapshot.take_pricing_snapshot` directly with explicit
send_alert injection. Same drift-alert behavior; no bot.py middleman.

Live verified after rebuild — bridge starts clean, all background
tasks (monitor, daily reporter, budget sweep, pricing snapshot)
running:
- 176/176 pytest pass (unchanged — cmd handlers themselves aren't
  unit tested, full integration covers the dispatch path)
- ruff clean

bot.py: 1,538 → 1,286 lines (-252). Total since start: 3,579 → 1,286
(-2,293, -64%).
