+++ telegram-bridge/telegram_handlers/__init__.py
@@ -7,6 +7,10 @@ pricing, usage, system, monitor toggles, …) into their own files
 under this package as their own bot.py-internal deps come unstuck.
 """
 
+from .cost import cmd_budget, cmd_pricing, cmd_usage
 from .today import _parse_by_flag, cmd_tasks, cmd_today, cmd_week
 
-__all__ = ["_parse_by_flag", "cmd_tasks", "cmd_today", "cmd_week"]
+__all__ = [
+    "_parse_by_flag", "cmd_tasks", "cmd_today", "cmd_week",
+    "cmd_budget", "cmd_pricing", "cmd_usage",
+]
